+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brickflow/i18n
 
+## 0.0.12
+
+### Patch Changes
+
+- Fix workspace cli translate
+
 ## 0.0.11
 
 ### Patch Changes